make WorkshopItem class interhits from WorkspaceFolderItem class. WorkspaceFolderItem interhits from TreeViewItem class.
Mod ID's type actually System.UInt64 (ulong)
