+++ DSTEd/UI/Components/ModInfoItem.cs
@@ -1,7 +1,11 @@
 ﻿using System.Windows.Controls;
 
 namespace DSTEd.UI.Components {
-    class ModInfoItem : WorkspaceFileItem {
-		public ModInfoItem(string FullPath, System.Windows.Input.MouseButtonEventHandler CustomPreviewCallback = null) : base(FullPath, CustomPreviewCallback) { }
+    class ModInfoItem : WorkspaceFileItem
+	{
+		public ModInfoItem(string FullPath) : base(FullPath)
+		{
+
+		}
 	}
 }
