+++ DSTEd/UI/Components/TextureItem.cs
@@ -1,7 +1,11 @@
 ﻿using System.Windows.Controls;
 
 namespace DSTEd.UI.Components {
-    class TextureItem : WorkspaceFileItem {
-		public TextureItem(string FullPath,System.Windows.Input.MouseButtonEventHandler CustomPreviewCallback = null) : base(FullPath, CustomPreviewCallback) { }
+    class TextureItem : WorkspaceFileItem
+	{
+		public TextureItem(string FullPath) : base(FullPath)
+		{
+
+		}
     }
 }
